requirements/common.txt
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.0
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-SQLAlchemy==2.2
Flask-Script==2.0.5
Flask-WTF==0.14.2
Jinja2==2.9.6
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
Pillow==4.2.1
SQLAlchemy==0.9.9
WTForms==2.1
Werkzeug==0.12.2
alembic==0.9.3
bleach==1.4.0
blinker==1.4
bs4==0.0.1
html5lib==1.0b3
itsdangerous==0.24
six==1.10.0
